change src_shared

+++ Buffer_shared.cpp
@@ -77,8 +77,8 @@ void FacetHistogramBuffer::Reset() {
 #endif
 }
 
-size_t FacetHistogramBuffer::GetMemSize() const {
-	size_t sum = 0;
+int FacetHistogramBuffer::GetMemSize() const {
+	int sum = 0;
 	sum += nbHitsHistogram.capacity() * sizeof(double);
 	sum += distanceHistogram.capacity() * sizeof(double);
 #ifdef MOLFLOW
